+++ lib/skein_rest_client.dart
@@ -5,7 +5,7 @@ import 'src/rest_config.dart';
 export 'src/exceptions.dart';
 export 'src/http/http.dart';
 export 'src/rest.dart';
-export 'src/rest_client.dart';
+export 'src/rest_client.dart' hide RestClientRegistry;
 export 'src/rest_config.dart';
 
 class Rest {