+++ lib/src/rest_client.dart
@@ -3,13 +3,18 @@ import 'dart:async';
 import 'package:async/async.dart';
 import 'package:skein_rest_client/skein_rest_client.dart';
 
+part 'rest_client_registry.dart';
+
 typedef DecoderFunction<T> = FutureOr<T> Function(dynamic value);
 typedef EncoderFunction<T> = FutureOr<dynamic> Function(T value);
 typedef AuthorizationBuilder = FutureOr<Authorization?> Function();
 
 abstract class RestClient {
 
-  late Uri uri;
+  Uri? _uri;
+  Uri get uri => _uri!;
+
+  void init(Uri uri) => _uri = uri;
 
   // MARK: Decoder
 
@@ -43,17 +48,27 @@ abstract class RestClient {
     return this;
   }
 
-  void init(Uri uri) => this.uri = uri;
+  // MARK: Stub
+
+  FutureOr? _stub;
+  RestClient stub<T>(FutureOr<T> stub) {
+    _stub = stub;
+    return this;
+  }
 
   // MARK: HTTP Methods
 
   CancelableOperation<T> post<T>([dynamic data]);
-
   CancelableOperation<T> patch<T>([dynamic data]);
-
-  CancelableOperation<T> get<T>();
-
+  CancelableOperation<T> get<T>() ;
   CancelableOperation<T> delete<T>([dynamic data]);
+
+  // MARK: HTTP methods to implement
+
+  CancelableOperation<T> doPost<T>([dynamic data]);
+  CancelableOperation<T> doPatch<T>([dynamic data]);
+  CancelableOperation<T> doGet<T>();
+  CancelableOperation<T> doDelete<T>([dynamic data]);
 
 }
 
@@ -79,18 +94,56 @@ mixin RestClientHelper on RestClient {
   }
 
   FutureOr<dynamic> encodeIfNeeded<T>(T value) {
-    if (_encoder == null) {
-      return value;
-    }
-    return _encoder!(value);
+    return _encoder != null ? _encoder!(value) : value;
   }
 
   FutureOr<T> decodeIfNeeded<T>(dynamic value) {
-    final aDecoder = _decoder as DecoderFunction<T>?;
-    if (aDecoder == null) {
-      return value;
-    }
-    return aDecoder(value);
+    return _decoder != null ? _decoder!(value) : value;
+  }
+
+  @override
+  CancelableOperation<T> post<T>([dynamic data]) {
+    final CancelableOperation<T> operation = _stub != null ? _wrap(stub: _stub) : doPost(data);
+    operation.onEnd(() => RestClientRegistry.reuse(this));
+    return operation;
+  }
+
+  @override
+  CancelableOperation<T> patch<T>([dynamic data]) {
+    final CancelableOperation<T> operation = _stub != null ? _wrap(stub: _stub) : doPatch(data);
+    operation.onEnd(() => RestClientRegistry.reuse(this));
+    return operation;
+  }
+
+  @override
+  CancelableOperation<T> get<T>() {
+    final CancelableOperation<T> operation = _stub != null ? _wrap(stub: _stub) : doGet();
+    operation.onEnd(() => RestClientRegistry.reuse(this));
+    return operation;
+  }
+
+  @override
+  CancelableOperation<T> delete<T>([dynamic data]) {
+    final CancelableOperation<T> operation = _stub != null ? _wrap(stub: _stub) : doDelete(data);
+    operation.onEnd(() => RestClientRegistry.reuse(this));
+    return operation;
+  }
+
+  CancelableOperation<T> _wrap<T>({required FutureOr stub}) {
+    return CancelableOperation.fromFuture(Future(() async {
+      return _decoder != null ? _decoder!(await _stub) : await _stub;
+    }));
+  }
+
+}
+
+extension on CancelableOperation {
+
+  CancelableOperation<R> onEnd<R>(FutureOr<R> Function() listener) {
+    return then((_) => listener(),
+      onError: (_, __) => listener(),
+      onCancel: () => listener(),
+    );
   }
 
 }