+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+
+* Adds support for stubbing responses via `stub()` function
+* Introduces `RestClientRegistry` internal class that manages `RestClient` instances.
+
 ## 0.0.1
 
 * Initial release, the next HTTP methods are supported: POST, GET, PATCH, DELETE.